# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi Kece",
        "description": "Dokumentasi endpoint lumen - swagger",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/kabupaten": {
            "get": {
                "tags": [
                    "Kabupaten"
                ],
                "summary": "Get List Kabupaten",
                "operationId": "getKabupaten",
                "responses": {
                    "200": {
                        "description": "Returns all area data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kabupaten": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "kabupaten_id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "provinsi": {
                                                        "properties": {
                                                            "provinsi_id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/kabupaten/{kabupaten_id}": {
            "get": {
                "tags": [
                    "Kabupaten"
                ],
                "summary": "Get Kabupaten Data By ID",
                "operationId": "/kabupaten/kabupatenId",
                "parameters": [
                    {
                        "name": "kabupaten_id",
                        "in": "path",
                        "description": "The id of kabupaten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show kabupaten data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kabupaten_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "provinsi": {
                                            "properties": {
                                                "provinsi_id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kabupaten Not Found"
                    }
                }
            }
        },
        "/api/create-kabupaten": {
            "post": {
                "tags": [
                    "Kabupaten"
                ],
                "summary": "Create New Kabupaten",
                "operationId": "createKabupaten",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateKabupatenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/CreateKabupatenResult"
                                        },
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/update-kabupaten/{kabupaten_id}": {
            "put": {
                "tags": [
                    "Kabupaten"
                ],
                "summary": "Update Kabupaten",
                "operationId": "updateKabupaten",
                "parameters": [
                    {
                        "name": "kabupaten_id",
                        "in": "path",
                        "description": "The id of kabupaten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateKabupatenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/UpdateKabupatenResult"
                                        },
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/delete-kabupaten/{kabupaten_id}": {
            "delete": {
                "tags": [
                    "Kabupaten"
                ],
                "summary": "Delete Kabupaten",
                "operationId": "deleteKabupaten",
                "parameters": [
                    {
                        "name": "kabupaten_id",
                        "in": "path",
                        "description": "The id of kabupaten",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateKabupatenRequest": {
                "title": "Create Kabupaten Schema",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "provinsi_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CreateKabupatenResult": {
                "title": "Response Create Kabupaten",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "kabupaten_id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "provinsi": {
                                "properties": {
                                    "provinsi_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "pbject"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateKabupatenRequest": {
                "title": "Create Kabupaten Schema",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "provinsi_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UpdateKabupatenResult": {
                "title": "Response Update Porvinsi",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "provinsi_id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}